% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_resource.R
\name{download_resource}
\alias{download_resource}
\title{Download a Resource File}
\usage{
download_resource(resource_url, download_folder)
}
\arguments{
\item{resource_url}{URL of the resource file to download.}

\item{download_folder}{Folder where the downloaded file should be saved.}
}
\value{
A character value containing the path to the downloaded file.
Returns \code{NA_character_} if the download fails.
}
\description{
Downloads a file from a supplied resource URL and saves it to a specified
folder.
}
\details{
The download folder is created automatically if it does not already exist.
If the file has already been downloaded, the existing file is returned
without downloading it again.
}
\examples{
\dontrun{
downloaded_file <- download_resource(
  resource_url = paste0(
    "https://digital.nhs.uk/example/",
    "community-services-data.zip"
  ),
  download_folder = "data/CSDS_downloads"
)

downloaded_file
}

}
