% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_resource_link.R
\name{get_resource_link}
\alias{get_resource_link}
\title{Get Resource Link}
\usage{
get_resource_link(
  publication_url,
  dataset_suffix = "/datasets",
  resource_text,
  period_pattern = "[a-z]+-[0-9]{4}/?$",
  match_period = TRUE,
  file_pattern = "\\\\.zip($|\\\\?)"
)
}
\arguments{
\item{publication_url}{URL of the publication page.}

\item{dataset_suffix}{Suffix appended to the publication URL to locate the
dataset page. Defaults to \code{"/datasets"}.}

\item{resource_text}{Text used to identify the required resource link.
Matching is case-insensitive.}

\item{period_pattern}{Regular expression used to extract the reporting
period from the publication URL. Defaults to
\code{"[a-z]+-[0-9]{4}/?$"}.}

\item{match_period}{Logical. If \code{TRUE}, the resource link text must contain
the reporting period extracted from \code{publication_url}.}

\item{file_pattern}{Regular expression used to identify the required file
type. Defaults to ZIP files using \code{"\\\\.zip($|\\\\?)"}.}
}
\value{
A character value containing the resource URL, or \code{NA_character_}
if no matching resource is found.
}
\description{
Retrieves a resource download link from a publication dataset page.
}
\examples{
\dontrun{
resource_url <- get_resource_link(
  publication_url = paste0(
    "https://digital.nhs.uk/data-and-information/publications/",
    "statistical/community-services-statistics-for-children-young-people-and-adults/",
    "june-2026"
  ),
  dataset_suffix = "/datasets",
  resource_text = "CSV Data \\\\(as ZIP\\\\)",
  period_pattern = "[a-z]+-[0-9]{4}/?$",
  match_period = TRUE,
  file_pattern = "\\\\.zip($|\\\\?)"
)

resource_url
}

}
