% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricengineR-package.R
\docType{package}
\name{metricengineR-package}
\alias{metricengineR}
\alias{metricengineR-package}
\title{metricengineR: Reusable Functions for Metric Engine Pipelines}
\description{
Provides reusable functions for metric engine workflows, including database extraction and loading, indicator processing, data standardisation, calculation helpers, output tidying, and supporting transformations used across analytical frameworks such as oversight, outcomes, and health inequalities reporting.
}
\author{
\strong{Maintainer}: Siti Hassan \email{siti.hassan@nhs.net}

Authors:
\itemize{
  \item Siti Hassan \email{siti.hassan@nhs.net}
}

}
\keyword{internal}
