% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_child_links.R
\name{get_child_links}
\alias{get_child_links}
\title{Get Child Publication Links}
\usage{
get_child_links(parent_url, publication_pattern)
}
\arguments{
\item{parent_url}{URL of the parent web page.}

\item{publication_pattern}{Regular expression used to identify publication
links.}
}
\value{
A character vector containing unique matching publication URLs.
}
\description{
Retrieves links from a parent web page that match a supplied publication
pattern and converts them to absolute URLs.
}
