% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_absolute_url.R
\name{make_absolute_url}
\alias{make_absolute_url}
\title{Make a URL Absolute}
\usage{
make_absolute_url(link, base_url)
}
\arguments{
\item{link}{Character vector containing URLs or relative links.}

\item{base_url}{Base URL used to resolve relative links.}
}
\value{
A character vector containing absolute URLs.
}
\description{
Converts a relative URL into an absolute URL using a supplied base URL.
Absolute URLs are returned unchanged.
}
